"Finished" maze.window.Window and starting to dive into maze.main.Main
I managed to draw a png image on the window and center it.
Maybe in the future, instead of trying to draw the maze from the maze
object, just transform it to a png and then let Main do it's thing? I
don't know, maybe it's harder to save an object as an image...

 On branch master
 Changes to be committed:
        new file:   res/maze.png
        modified:   src/maze/main/Main.java
        modified:   src/maze/window/Window.java