Added the "nodes" file whenever you created a maze, it's simply a binary file that stores every node (with it's x and y coordinates). I've done it because I thought it would have been an easier and faster was to add nodes in the data structure since we saw them once going through all the maze, why go look for them again while we can simply just store them and use them later.
But it turn out that the "searching for nodes" part wasn't the one that took that long, it was just the "linkNodes" that was written really poorly. So I did it at the same time. The time to create the nodes decreased a lot (for example, from 18 seconds to 2 seconds on braid2k).